Add default layouts to configuration
All pages will use the "post" layout by default. Pages in the
`brewing/` directory use the "brew" layout by default. This removes the
requirement to remember to add `layout: post` to the frontmatter for
each new post/page. Only pages that break from the norm need to define a
layout.